requirements.txt
# Exchange API
ccxt>=4.4.0

# Data & Analysis
pandas>=2.2.0
numpy>=1.26.0
ta>=0.11.0
scipy>=1.12.0

# WebSocket
websockets>=13.0

# Schema Validation
pydantic>=2.6.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0

# Scheduling
APScheduler>=3.10.0

# Logging
colorlog>=6.8.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0

# Backtest visualization (optional)
matplotlib>=3.8.0

# Supabase (database layer)
supabase>=2.0.0
